+++ packages/ui/src/hooks/index.ts
@@ -24,6 +24,7 @@ import useGasTopUpRequired from './useGasTopUpRequired.js'
 import useHyperliquidUsdcBalance from './useHyperliquidUsdcBalance.js'
 import useEOADetection from './useEOADetection.js'
 import useTransactionCount from './useTransactionCount.js'
+import useTronBalance from './useTronBalance.js'
 
 export {
   useMounted,
@@ -51,5 +52,6 @@ export {
   useGasTopUpRequired,
   useHyperliquidUsdcBalance,
   useEOADetection,
-  useTransactionCount
+  useTransactionCount,
+  useTronBalance
 }